requirements.txt
streamlit
numpy
tensorflow
pillow
matplotlib
seaborn
python-dotenv
openai
